chore: put everything under the module of this project (#20)
