+++ lib/leaf_api.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require_relative 'leaf_api/entity'
+require_relative 'leaf_api/service'
