+++ lib/leaf_api/entity.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require_relative 'entity/trip'
+require_relative 'entity/location'
+require_relative 'entity/travel_strategy'
