+++ lib/leaf_api/service.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require_relative 'service/api_errors'
+require_relative 'service/nthu_api'
+require_relative 'service/llama_api'
+require_relative 'service/nominatim_api'
+require_relative 'service/google_maps_api'
