+++ spec/api/google_maps_api_spec.rb
@@ -2,9 +2,6 @@
 
 require_relative '../spec_helper'
 
-require_relative '../../lib/services/google_maps_api'
-require_relative '../../lib/services/api_errors'
-
 describe 'Test Google API library' do
   VCR.configure do |c|
     c.cassette_library_dir = CASSETTES_FOLDER
@@ -27,16 +24,16 @@ describe 'Test Google API library' do
   describe 'API Authentication Failed' do
     it 'Raise errors when provided with incorrect token on distance matrix.' do
       _(proc do
-        GoogleMapsAPI.new(BAD_SECRETS['GOOGLE_TOKEN'])
+        LeafAPI::Service::GoogleMapsAPI.new(BAD_SECRETS['GOOGLE_TOKEN'])
                      .distance_matrix('光明里 300, Hsinchu City, East District', '24.8022,120.9901', 'walking')
-      end).must_raise HTTPError
+      end).must_raise LeafAPI::Service::HTTPError
     end
 
     it 'Raise errors when provided with incorrect token on distance matrix.' do
       _(proc do
-        GoogleMapsAPI.new(BAD_SECRETS['GOOGLE_TOKEN'])
+        LeafAPI::Service::GoogleMapsAPI.new(BAD_SECRETS['GOOGLE_TOKEN'])
                      .geocoding('光明里 300, Hsinchu City, East District')
-      end).must_raise HTTPError
+      end).must_raise LeafAPI::Service::HTTPError
     end
   end
 
@@ -44,8 +41,8 @@ describe 'Test Google API library' do
     it 'Receive correct data for distance matrix.' do
       correct_response = YAML.safe_load_file('spec/fixtures/google_maps_distance_matrix-results.yaml')
 
-      payload = GoogleMapsAPI.new(CORRECT_SECRETS['GOOGLE_TOKEN'])
-                             .distance_matrix('光明里 300, Hsinchu City, East District', '24.8022,120.9901', 'walking')
+      payload = LeafAPI::Service::GoogleMapsAPI.new(CORRECT_SECRETS['GOOGLE_TOKEN'])
+                                               .distance_matrix('光明里 300, Hsinchu City, East District', '24.8022,120.9901', 'walking')
       _(payload['destination_addresses'][0]).must_equal correct_response['destination_addresses'][0]
       _(payload['origin_addresses'][0]).must_equal correct_response['origin_addresses'][0]
       _(payload['rows'][0]['elements'][0]).wont_be_nil
@@ -55,8 +52,8 @@ describe 'Test Google API library' do
 
     it 'Receive correct data for geocoding.' do
       correct_response = YAML.safe_load_file('spec/fixtures/google_maps_geocoding-results.yaml')
-      payload = GoogleMapsAPI.new(CORRECT_SECRETS['GOOGLE_TOKEN'])
-                             .geocoding('光明里 300, Hsinchu City, East District')
+      payload = LeafAPI::Service::GoogleMapsAPI.new(CORRECT_SECRETS['GOOGLE_TOKEN'])
+                                               .geocoding('光明里 300, Hsinchu City, East District')
       _(payload['results'][0]['formatted_address'])
         .must_equal correct_response['results'][0]['formatted_address']
       _(payload['results'][0]['geometry']['location'])
