+++ spec/api/nominatim_api_spec.rb
@@ -2,9 +2,6 @@
 
 require_relative '../spec_helper'
 
-require_relative '../../lib/services/nominatim_api'
-require_relative '../../lib/services/api_errors'
-
 describe 'Test Nominatim API library' do
   VCR.configure do |c|
     c.cassette_library_dir = CASSETTES_FOLDER
@@ -24,7 +21,7 @@ describe 'Test Nominatim API library' do
   describe 'API Search Succeed' do
     it 'Receive correct data for NTHU search.' do
       correct_response = YAML.safe_load_file('spec/fixtures/nominatim_serach_nthu-results.yaml')
-      nominatim_api = NominatimAPI.new
+      nominatim_api = LeafAPI::Service::NominatimAPI.new
       payload = nominatim_api.search('清華大學')
       _(payload[0]['place_id']).must_equal correct_response[0]['place_id']
       _(payload[0]['name']).must_equal correct_response[0]['name']
