+++ spec/api/nthu_api_spec.rb
@@ -2,9 +2,6 @@
 
 require_relative '../spec_helper'
 
-require_relative '../../lib/services/nthu_api'
-require_relative '../../lib/services/api_errors'
-
 describe 'Test NTHUAPI library' do
   VCR.configure do |c|
     c.cassette_library_dir = CASSETTES_FOLDER
@@ -24,7 +21,7 @@ describe 'Test NTHUAPI library' do
   describe 'API Authentication Suceed' do
     it 'Receive correct data.' do
       correct_response = [YAML.safe_load_file('spec/fixtures/nthuapi_bus_schedule-results.yaml')].freeze
-      payload = NTHUAPI.new.bus_schedule('北校門口', 'main', 'up', 'all')[0]
+      payload = LeafAPI::Service::NTHUAPI.new.bus_schedule('北校門口', 'main', 'up', 'all')[0]
       _(payload).wont_be_nil
       formatted_payload = payload.is_a?(Hash) ? [payload] : payload
       _(formatted_payload).must_equal correct_response
