+++ spec/spec_helper.rb
@@ -11,6 +11,8 @@ require 'minitest/rg'
 require 'vcr'
 require 'webmock'
 
+require_relative '../lib/leaf_api'
+
 CORRECT_SECRETS = YAML.safe_load_file('config/secrets.yaml')
 BAD_SECRETS = YAML.safe_load_file('config/secrets.yaml.example')
